refactor(rust): Split hash.rs into modular structure
